+++ Panels/Header.js
@@ -16,7 +16,6 @@ import React from 'react';
 import $L from '../internal/$L';
 import Button from '../Button';
 import Heading from '../Heading';
-import Skinnable from '../Skinnable';
 import {useScrollPosition} from '../useScroll/useScrollPosition';
 import WindowEventable from '../internal/WindowEventable';
 
@@ -451,8 +450,7 @@ const HeaderBase = kind({
 		noCloseButton: preferPropOverContext('noCloseButton'),
 		onBack: preferPropOverContext('onBack'),
 		onClose: preferPropOverContext('onClose'),
-		direction: ({title, subtitle}) => isRtlText(title) || isRtlText(subtitle) ? 'rtl' : 'ltr',
-		line: ({css, type}) => ((type === 'compact') && <Cell shrink component="hr" className={css.line} />)
+		direction: ({title, subtitle}) => isRtlText(title) || isRtlText(subtitle) ? 'rtl' : 'ltr'
 	},
 
 	render: ({
@@ -466,7 +464,6 @@ const HeaderBase = kind({
 		css,
 		direction,
 		hover,
-		line,
 		marqueeOn,
 		noBackButton,
 		noCloseButton,
@@ -560,7 +557,6 @@ const HeaderBase = kind({
 					</Cell>
 				</Row>
 				{children ? <nav className={css.slotBelow}>{children}</nav> : null}
-				{line}
 			</header>
 		);
 	}
@@ -605,7 +601,6 @@ const HeaderMeasurementDecorator = (Wrapped) => {
 
 const HeaderDecorator = compose(
 	Slottable({slots: ['title', 'subtitle', 'slotAbove', 'slotAfter', 'slotBefore']}),
-	Skinnable,
 	CollapsingHeaderDecorator,
 	HeaderMeasurementDecorator,
 	Toggleable({prop: 'hover', activate: 'onShowBack', deactivate: 'onHideBack', toggle: null}),