+++ CHANGELOG.md
@@ -13,11 +13,17 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/FixedPopupPanels` `width` prop, which now includes "half" to support larger content
 - `sandstone/MediaPlayer` submodule which provides `MediaControls`, `MediaSlider`, and `Times` components for use in custom media player components.
 - `sandstone/WizardPanels` props `current` and `total`
+- `sandstone/WizardPanels` prop `onBack` to allow developers to handle back button presses
 
 ### Changed
 
 - `sandstone/Scroller` and `sandstone/VirtualList` to adjust padding area
 - `sandstone/Scroller` and `sandstone/VirtualList` clickable scrollbar area and condition
+- `sandstone/WizardPanels` to automatically handle back key when `noPrevButton` is not set
+
+### Fixed
+
+- `sandstone/Panels.Header` to match latest designs
 
 ## [1.0.0-beta.2] - 2020-05-04